add basic client and move submission logic from executor
